:sparkles:feat: Firebase -> RoomDB, 패키지명 변경, HomeFragment 리스트, spinner, 데이터 연동 및 수정
