--- SeoulData/app/src/main/java/com/example/seouldata/SplashActivity.kt
@@ -1,34 +1,0 @@
-package com.example.seouldata
-
-import android.content.Intent
-import android.os.Bundle
-import android.os.Handler
-import android.os.Looper
-import androidx.appcompat.app.AppCompatActivity
-import com.example.seouldata.util.UserExpManager
-import com.google.firebase.auth.FirebaseAuth
-import java.text.SimpleDateFormat
-import java.util.Date
-import java.util.Locale
-
-import androidx.core.splashscreen.SplashScreen;
-
-class SplashActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-
-
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_splash)
-
-Handler(Looper.getMainLooper()).postDelayed({
-            val user = FirebaseAuth.getInstance().currentUser
-            if (user != null) {
-                startActivity(Intent(this, MainActivity::class.java))
-            } else {
-                startActivity(Intent(this, LoginActivity::class.java))
-            }
-            finish() // 스플래시 종료
-        }, 1000) //  1초간 스플래시 화면 유지
-    }
-
-}
