--- SeoulData/app/src/main/java/com/example/seouldata/dto/InventoryItem.kt
@@ -1,6 +1,0 @@
-package com.example.seouldata.dto
-
-data class InventoryItem(
-    val name: String,
-    val drawableResId: Int
-)
