--- SeoulData/app/src/main/java/com/example/seouldata/ui/home/HomeViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.seouldata.ui.home
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class HomeViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "SportsFacilityApp"
-    }
-    val text: LiveData<String> = _text
-}
