--- SeoulData/app/src/main/java/com/example/seouldata/ui/notifications/NotificationsViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.seouldata.ui.notifications
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class NotificationsViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
-    }
-    val text: LiveData<String> = _text
-}
