+++ SeoulData/app/src/main/java/com/seouldata/sport/SplashActivity.kt
@@ -1,0 +1,28 @@
+package com.seouldata.sport
+
+import android.content.Intent
+import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
+import androidx.appcompat.app.AppCompatActivity
+import com.google.firebase.auth.FirebaseAuth
+
+class SplashActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+
+
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_splash)
+
+Handler(Looper.getMainLooper()).postDelayed({
+            val user = FirebaseAuth.getInstance().currentUser
+            if (user != null) {
+                startActivity(Intent(this, MainActivity::class.java))
+            } else {
+                startActivity(Intent(this, LoginActivity::class.java))
+            }
+            finish() // 스플래시 종료
+        }, 1000) //  1초간 스플래시 화면 유지
+    }
+
+}
