+++ SeoulData/app/src/main/java/com/seouldata/sport/dto/InventoryItem.kt
@@ -1,0 +1,6 @@
+package com.seouldata.sport.dto
+
+data class InventoryItem(
+    val name: String,
+    val drawableResId: Int
+)
