+++ SeoulData/app/src/main/java/com/seouldata/sport/ui/dashboard/DashboardViewModel.kt
@@ -1,0 +1,13 @@
+package com.seouldata.sport.ui.dashboard
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class DashboardViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "This is dashboard Fragment"
+    }
+    val text: LiveData<String> = _text
+}
