+++ SeoulData/app/src/main/java/com/seouldata/sport/ui/home/HomeViewModel.kt
@@ -1,0 +1,13 @@
+package com.seouldata.sport.ui.home
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+
+class HomeViewModel : ViewModel() {
+
+    private val _text = MutableLiveData<String>().apply {
+        value = "SportsFacilityApp"
+    }
+    val text: LiveData<String> = _text
+}
